requirements.txt
asgiref==3.7.2
boto3==1.28.76
botocore==1.31.76
Django==4.2.6
django-appconf==1.0.5
django-compressor==4.4
django-environ==0.11.2
django-ses==3.5.0
jmespath==1.0.1
Pillow==10.1.0
python-dateutil==2.8.2
pytz==2023.3.post1
rcssmin==1.1.1
rjsmin==1.2.1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.7
whitenoise==6.6.0
